Fix white space or other non-word characters ignoring delimiters
Delimiters (including new-lines) that existed between words
were ignored if there was white space or other non-word characters
before them.

This meant (for example) that trailing space at the end of a line
would cause the new-line to be ignored as a delimiter.

Add tests - all of them failed before this fix.

---

This also refactors the main loop into two functions,
so capitalizing between delimiters is handled in an outer loop.

This simplifies logic elsewhere.