Refactor to optionally use an external script for `titlecase-command`
To be more friendly with prior art in this space (see
https://github.com/melpa/melpa/pull/7852), I've refactored this titlecase
package to enable the user to specify a `titlecase-command`, which can be a
function, string, or list of strings.  It defaults to this package's internal
implementation of title-casing, but can be set to an external script or even
another function.

I've also adapted the `titlecase-region` function to use a temporary buffer,
like Jason Blevin's implementation, and attempted to make undoes easier to deal
with by moving to the end of the sentence or line titlecased (addressing issue